Move HasExtensionOptionsTx to authtx?
