+++ x/auth/ante/ext.go
@@ -1,15 +1,10 @@
 package ante
 
 import (
-	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
 	"github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
+	authtx "github.com/cosmos/cosmos-sdk/x/auth/tx"
 )
-
-type HasExtensionOptionsTx interface {
-	GetExtensionOptions() []*codectypes.Any
-	GetNonCriticalExtensionOptions() []*codectypes.Any
-}
 
 // RejectExtensionOptionsDecorator is an AnteDecorator that rejects all extension
 // options which can optionally be included in protobuf transactions. Users that
@@ -26,7 +21,7 @@ var _ types.AnteDecorator = RejectExtensionOptionsDecorator{}
 
 // AnteHandle implements the AnteDecorator.AnteHandle method
 func (r RejectExtensionOptionsDecorator) AnteHandle(ctx types.Context, tx types.Tx, simulate bool, next types.AnteHandler) (newCtx types.Context, err error) {
-	if hasExtOptsTx, ok := tx.(HasExtensionOptionsTx); ok {
+	if hasExtOptsTx, ok := tx.(authtx.HasExtensionOptionsTx); ok {
 		if len(hasExtOptsTx.GetExtensionOptions()) != 0 {
 			return ctx, sdkerrors.ErrUnknownExtensionOptions
 		}
