+++ x/auth/tx/builder.go
@@ -10,9 +10,13 @@ import (
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 	"github.com/cosmos/cosmos-sdk/types/tx"
 	"github.com/cosmos/cosmos-sdk/types/tx/signing"
-	"github.com/cosmos/cosmos-sdk/x/auth/ante"
 	authsigning "github.com/cosmos/cosmos-sdk/x/auth/signing"
 )
+
+type HasExtensionOptionsTx interface {
+	GetExtensionOptions() []*codectypes.Any
+	GetNonCriticalExtensionOptions() []*codectypes.Any
+}
 
 // wrapper is a wrapper around the tx.Tx proto.Message which retain the raw
 // body and auth_info bytes.
@@ -31,10 +35,10 @@ type wrapper struct {
 }
 
 var (
-	_ authsigning.Tx             = &wrapper{}
-	_ client.TxBuilder           = &wrapper{}
-	_ ante.HasExtensionOptionsTx = &wrapper{}
-	_ ExtensionOptionsTxBuilder  = &wrapper{}
+	_ authsigning.Tx            = &wrapper{}
+	_ client.TxBuilder          = &wrapper{}
+	_ HasExtensionOptionsTx     = &wrapper{}
+	_ ExtensionOptionsTxBuilder = &wrapper{}
 )
 
 // ExtensionOptionsTxBuilder defines a TxBuilder that can also set extensions.
